requirements.txt
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
fairscale==0.4.4
filelock==3.11.0
fonttools==4.39.3
huggingface-hub==0.13.3
importlib-resources==5.12.0
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
Pillow==9.4.0
pycocoevalcap==1.2
pycocotools==2.0.6
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sacremoses==0.0.53
six==1.16.0
timm==0.4.12
tokenizers==0.10.3
torch==1.10.0
torchvision==0.11.1
tqdm==4.65.0
transformers==4.15.0
typing_extensions==4.5.0
urllib3==1.26.15
zipp==3.15.0